Be more defensive about purple_conversation_get_data
When we're looking up information from
purple_conversation_get_data(conv, ...), it may happen that the result
is NULL because pidgin has emitted conversation-switched before emitting
conversation-created, so we try to switch to a conversation that the
conversation-created callback hasn't yet filled in the data for.  This
causes a crash.  Be more defensive about this.